restructure unit tests in common module
