+++ noisy-simulator/src/density_matrix_simulator.rs
@@ -9,7 +9,7 @@ mod tests;
 
 use crate::{
     handle_error, instrument::Instrument, kernel::apply_kernel, operation::Operation,
-    ComplexVector, Error, SquareMatrix, TOLERANCE,
+    ComplexVector, Error, NoisySimulator, SquareMatrix, TOLERANCE,
 };
 use num_complex::Complex;
 
@@ -166,9 +166,11 @@ pub struct DensityMatrixSimulator {
     dimension: usize,
 }
 
-impl DensityMatrixSimulator {
+impl NoisySimulator for DensityMatrixSimulator {
+    type State = DensityMatrix;
+
     /// Creates a new `DensityMatrixSimulator`.
-    pub fn new(number_of_qubits: usize) -> Self {
+    fn new(number_of_qubits: usize) -> Self {
         let density_matrix = DensityMatrix::new(number_of_qubits);
         let dimension = density_matrix.dimension();
         Self {
@@ -178,11 +180,7 @@ impl DensityMatrixSimulator {
     }
 
     /// Apply an operation to the given qubit ids.
-    pub fn apply_operation(
-        &mut self,
-        operation: &Operation,
-        qubits: &[usize],
-    ) -> Result<(), Error> {
+    fn apply_operation(&mut self, operation: &Operation, qubits: &[usize]) -> Result<(), Error> {
         self.state
             .as_mut()?
             .apply_operation_matrix(operation.matrix(), qubits)?;
@@ -193,11 +191,7 @@ impl DensityMatrixSimulator {
     }
 
     /// Apply non selective evolution to the given qubit ids.
-    pub fn apply_instrument(
-        &mut self,
-        instrument: &Instrument,
-        qubits: &[usize],
-    ) -> Result<(), Error> {
+    fn apply_instrument(&mut self, instrument: &Instrument, qubits: &[usize]) -> Result<(), Error> {
         self.state
             .as_mut()?
             .apply_operation_matrix(instrument.non_selective_operation_matrix(), qubits)?;
@@ -211,7 +205,7 @@ impl DensityMatrixSimulator {
     /// Returns the index of the observed outcome.
     ///
     /// Use this method to perform measurements on the quantum system.
-    pub fn sample_instrument(
+    fn sample_instrument(
         &mut self,
         instrument: &Instrument,
         qubits: &[usize],
@@ -221,7 +215,7 @@ impl DensityMatrixSimulator {
 
     /// Performs selective evolution under the given instrument.
     /// Returns the index of the observed outcome.
-    pub fn sample_instrument_with_distribution(
+    fn sample_instrument_with_distribution(
         &mut self,
         instrument: &Instrument,
         qubits: &[usize],
@@ -282,12 +276,12 @@ impl DensityMatrixSimulator {
     }
 
     /// Returns the `DensityMatrix` if the simulator is in a valid state.
-    pub fn state(&self) -> Result<&DensityMatrix, &Error> {
+    fn state(&self) -> Result<&DensityMatrix, &Error> {
         self.state.as_ref()
     }
 
     /// Set state of the quantum system.
-    pub fn set_state(&mut self, new_state: DensityMatrix) -> Result<(), Error> {
+    fn set_state(&mut self, new_state: DensityMatrix) -> Result<(), Error> {
         if self.dimension != new_state.dimension() {
             return Err(Error::InvalidState(format!(
                 "the provided state should have the same dimensions as the quantum system's state, {} != {}",
@@ -311,12 +305,12 @@ impl DensityMatrixSimulator {
     /// Return theoretical change in trace due to operations that have been applied so far
     /// In reality, the density matrix is always renormalized after instruments/operations
     /// have been applied.
-    pub fn trace_change(&self) -> Result<f64, Error> {
+    fn trace_change(&self) -> Result<f64, Error> {
         Ok(self.state.as_ref()?.trace_change())
     }
 
     /// Set the trace of the quantum system.
-    pub fn set_trace(&mut self, trace: f64) -> Result<(), Error> {
+    fn set_trace(&mut self, trace: f64) -> Result<(), Error> {
         if trace < TOLERANCE || (trace - 1.) > TOLERANCE {
             return Err(Error::NotNormalized(trace));
         }
